daemon: move compileBase to new file, datapath.go
This splits up the code into files that are more logically organized.

Signed-off by: Ian Vernon <ian@cilium.io>
