+++ daemon/daemon.go
@@ -30,16 +30,12 @@ import (
 	health "github.com/cilium/cilium/cilium-health/launch"
 	"github.com/cilium/cilium/common"
 	"github.com/cilium/cilium/pkg/bpf"
-	"github.com/cilium/cilium/pkg/cgroups"
 	"github.com/cilium/cilium/pkg/clustermesh"
-	"github.com/cilium/cilium/pkg/command/exec"
 	"github.com/cilium/cilium/pkg/completion"
 	"github.com/cilium/cilium/pkg/controller"
 	"github.com/cilium/cilium/pkg/counter"
 	"github.com/cilium/cilium/pkg/datapath"
-	"github.com/cilium/cilium/pkg/datapath/alignchecker"
 	bpfIPCache "github.com/cilium/cilium/pkg/datapath/ipcache"
-	"github.com/cilium/cilium/pkg/datapath/iptables"
 	"github.com/cilium/cilium/pkg/datapath/linux/ipsec"
 	"github.com/cilium/cilium/pkg/datapath/loader"
 	"github.com/cilium/cilium/pkg/datapath/prefilter"
@@ -412,173 +408,6 @@ func (d *Daemon) setHostAddresses() error {
 // of BPF programs.
 func (d *Daemon) GetCompilationLock() *lock.RWMutex {
 	return d.compilationMutex
-}
-
-func (d *Daemon) compileBase() error {
-	var args []string
-	var mode string
-	var ret error
-
-	args = make([]string, initArgMax)
-
-	// Lock so that endpoints cannot be built while we are compile base programs.
-	d.compilationMutex.Lock()
-	defer d.compilationMutex.Unlock()
-
-	if err := d.writeNetdevHeader("./"); err != nil {
-		log.WithError(err).Warn("Unable to write netdev header")
-		return err
-	}
-	loader.Init(d.datapath, &d.nodeDiscovery.LocalConfig)
-
-	scopedLog := log.WithField(logfields.XDPDevice, option.Config.DevicePreFilter)
-	if option.Config.DevicePreFilter != "undefined" {
-		if err := prefilter.ProbePreFilter(option.Config.DevicePreFilter, option.Config.ModePreFilter); err != nil {
-			scopedLog.WithError(err).Warn("Turning off prefilter")
-			option.Config.DevicePreFilter = "undefined"
-		}
-	}
-	if option.Config.DevicePreFilter != "undefined" {
-		if d.preFilter, ret = prefilter.NewPreFilter(); ret != nil {
-			scopedLog.WithError(ret).Warn("Unable to init prefilter")
-			return ret
-		}
-
-		if err := d.writePreFilterHeader("./"); err != nil {
-			scopedLog.WithError(err).Warn("Unable to write prefilter header")
-			return err
-		}
-
-		args[initArgDevicePreFilter] = option.Config.DevicePreFilter
-		args[initArgModePreFilter] = option.Config.ModePreFilter
-	}
-
-	args[initArgLib] = option.Config.BpfDir
-	args[initArgRundir] = option.Config.StateDir
-	args[initArgCgroupRoot] = cgroups.GetCgroupRoot()
-	args[initArgBpffsRoot] = bpf.GetMapRoot()
-
-	if option.Config.EnableIPv4 {
-		args[initArgIPv4NodeIP] = node.GetInternalIPv4().String()
-	} else {
-		args[initArgIPv4NodeIP] = "<nil>"
-	}
-
-	if option.Config.EnableIPv6 {
-		args[initArgIPv6NodeIP] = node.GetIPv6().String()
-	} else {
-		args[initArgIPv6NodeIP] = "<nil>"
-	}
-
-	args[initArgMTU] = fmt.Sprintf("%d", d.mtuConfig.GetDeviceMTU())
-
-	if option.Config.EnableIPSec {
-		args[initArgIPSec] = "true"
-	} else {
-		args[initArgIPSec] = "false"
-	}
-
-	if !option.Config.InstallIptRules && option.Config.Masquerade {
-		args[initArgMasquerade] = "true"
-	} else {
-		args[initArgMasquerade] = "false"
-	}
-
-	if option.Config.EnableHostReachableServices {
-		args[initArgHostReachableServices] = "true"
-	} else {
-		args[initArgHostReachableServices] = "false"
-	}
-
-	if option.Config.EncryptInterface != "" {
-		args[initArgEncryptInterface] = option.Config.EncryptInterface
-	}
-
-	if option.Config.Device != "undefined" {
-		_, err := netlink.LinkByName(option.Config.Device)
-		if err != nil {
-			log.WithError(err).WithField("device", option.Config.Device).Warn("Link does not exist")
-			return err
-		}
-
-		if option.Config.IsLBEnabled() {
-			if option.Config.Device != option.Config.LBInterface {
-				//FIXME: allow different interfaces
-				return fmt.Errorf("Unable to have an interface for LB mode different than snooping interface")
-			}
-			if err := d.setHostAddresses(); err != nil {
-				return err
-			}
-			mode = "lb"
-		} else {
-			if option.Config.DatapathMode == option.DatapathModeIpvlan {
-				mode = "ipvlan"
-			} else {
-				mode = "direct"
-			}
-		}
-
-		args[initArgMode] = mode
-		args[initArgDevice] = option.Config.Device
-	} else {
-		if option.Config.IsLBEnabled() && strings.ToLower(option.Config.Tunnel) != "disabled" {
-			//FIXME: allow LBMode in tunnel
-			return fmt.Errorf("Unable to run LB mode with tunnel mode")
-		}
-
-		args[initArgMode] = option.Config.Tunnel
-
-		if option.Config.IsFlannelMasterDeviceSet() {
-			args[initArgMode] = "flannel"
-			args[initArgDevice] = option.Config.FlannelMasterDevice
-		}
-	}
-
-	prog := filepath.Join(option.Config.BpfDir, "init.sh")
-	ctx, cancel := context.WithTimeout(context.Background(), defaults.ExecTimeout)
-	defer cancel()
-	cmd := exec.CommandContext(ctx, prog, args...)
-	cmd.Env = bpf.Environment()
-	if _, err := cmd.CombinedOutput(log, true); err != nil {
-		return err
-	}
-
-	if canDisableDwarfRelocations {
-		// Validate alignments of C and Go equivalent structs
-		if err := alignchecker.CheckStructAlignments(defaults.AlignCheckerName); err != nil {
-			log.WithError(err).Fatal("C and Go structs alignment check failed")
-		}
-	} else {
-		log.Warning("Cannot check matching of C and Go common struct alignments due to old LLVM/clang version")
-	}
-
-	if !option.Config.IsFlannelMasterDeviceSet() {
-		d.ipam.ReserveLocalRoutes()
-	}
-
-	if err := d.datapath.Node().NodeConfigurationChanged(d.nodeDiscovery.LocalConfig); err != nil {
-		return err
-	}
-
-	iptablesManager := iptables.IptablesManager{}
-	iptablesManager.Init()
-	// Always remove masquerade rule and then re-add it if required
-	iptablesManager.RemoveRules()
-	if option.Config.InstallIptRules {
-		if err := iptablesManager.InstallRules(option.Config.HostDevice); err != nil {
-			return err
-		}
-	}
-	// Reinstall proxy rules for any running proxies
-	if d.l7Proxy != nil {
-		d.l7Proxy.ReinstallRules()
-	}
-
-	log.Info("Setting sysctl net.core.bpf_jit_enable=1")
-	log.Info("Setting sysctl net.ipv4.conf.all.rp_filter=0")
-	log.Info("Setting sysctl net.ipv6.conf.all.disable_ipv6=0")
-
-	return nil
 }
 
 // initMaps opens all BPF maps (and creates them if they do not exist). This
